refactor(auth): reuse generic oauth provider type
